backend/requirements.txt
# Core FastAPI & Server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database - MongoDB
motor==3.6.0
pymongo==4.9.1

# Data Validation & Serialization
pydantic[email]==2.10.3
email-validator
pydantic-settings==2.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pycryptodome==3.21.0

# File Upload & Processing
python-multipart==0.0.20
Pillow==11.0.0

# Environment Variables
python-dotenv==1.0.1

# QR Code Generation
qrcode[pil]==8.0

# PDF Generation
reportlab==4.2.5
fpdf2==2.8.1

# HTTP Client
httpx==0.28.1

# Task Queue (Optional - if using background tasks)
celery[redis]==5.4.0
redis==5.2.0
kombu==5.4.2
